refactor samples to have different package names so whole project can build
