+++ sample/hilt-di/build.gradle
@@ -9,7 +9,7 @@ android {
     compileSdk compile_sdk
 
     defaultConfig {
-        applicationId "com.plusmobileapps.savedstateflow"
+        applicationId "com.plusmobileapps.savedstateflowhilt"
         minSdk min_sdk
         targetSdk target_sdk
         versionCode 1
