+++ sample/hilt-di/src/main/java/com/plusmobileapps/savedstateflowhilt/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.plusmobileapps.savedstateflow
+package com.plusmobileapps.savedstateflowhilt
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -17,12 +17,14 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.plusmobileapps.savedstateflow.ui.theme.SavedStateFlowTheme
+import com.plusmobileapps.savedstateflowhilt.ui.theme.SavedStateFlowTheme
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class MainActivity : ComponentActivity() {
-    private val viewModel: MainViewModel by viewModels {
-        MainViewModelFactory(this)
-    }
+
+    private val viewModel: MainViewModel by viewModels()
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
