+++ sample/hilt-di/src/main/java/com/plusmobileapps/savedstateflowhilt/Modules.kt
@@ -1,4 +1,4 @@
-package com.plusmobileapps.savedstateflow
+package com.plusmobileapps.savedstateflowhilt
 
 import dagger.Binds
 import dagger.Module
