+++ sample/hilt-di/src/main/java/com/plusmobileapps/savedstateflowhilt/Repository.kt
@@ -1,14 +1,17 @@
-package com.plusmobileapps.savedstateflow
+package com.plusmobileapps.savedstateflowhilt
 
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
+import javax.inject.Inject
+import javax.inject.Singleton
 
 interface NewsDataSource {
     fun fetchQuery(query: String): Flow<List<String>>
 }
 
-object NewsRepository : NewsDataSource {
+@Singleton
+class NewsRepository @Inject constructor() : NewsDataSource {
     override fun fetchQuery(query: String): Flow<List<String>> = flow {
         delay(2000L)
         if (query.isEmpty()) {
