+++ sample/hilt-di/src/main/java/com/plusmobileapps/savedstateflowhilt/assisted/AssistedFragment.kt
@@ -1,4 +1,4 @@
-package com.plusmobileapps.savedstateflow.assisted
+package com.plusmobileapps.savedstateflowhilt.assisted
 
 import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
