+++ sample/hilt-di/src/main/java/com/plusmobileapps/savedstateflowhilt/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.plusmobileapps.savedstateflow.ui.theme
+package com.plusmobileapps.savedstateflowhilt.ui.theme
 
 import androidx.compose.ui.graphics.Color
