+++ sample/hilt-di/src/main/java/com/plusmobileapps/savedstateflowhilt/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.plusmobileapps.savedstateflow.ui.theme
+package com.plusmobileapps.savedstateflowhilt.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
