+++ sample/hilt-di/src/main/java/com/plusmobileapps/savedstateflowhilt/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.plusmobileapps.savedstateflow.ui.theme
+package com.plusmobileapps.savedstateflowhilt.ui.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
