+++ sample/manual-di/src/main/java/com/plusmobileapps/savedstateflowmanual/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.plusmobileapps.savedstateflow.ui.theme
+package com.plusmobileapps.savedstateflowmanual.ui.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
